Move image processing out of the requestedCompleted() callback; I have been getting a segmentation fault in the first function I call from the requestCompleted() callback, have raised this with Raspberry Pi (whose branch of libcamera one has to use on a Pi 5 since Pi 5 support is not yet in the main libcamera), see https://github.com/raspberrypi/libcamera/issues/221.
In case this was down to occupying the requestCompleted() callback, or the DMA buffer, for too long I though it would be sensible to move all my processing out of that function (since I end up taking a copy of the buffer anyway to queue it to the video stream encoder this doesn't really _cost_ anything more).

It doesn't help, unfortunately, the segmentation fault still occurs, but this solution is neater anyway so let's stick with it.