Finally, fixed the issue with the hls.js client not starting unless it was started at exactly the same time as the served stream:   One needs to tell the FFmpeg `libx264` video codec to `-tune zerolatency`.
Why? I couldn't tell you.  It has taken me nearly a week of trying everything I could try before I found this.  With this single option added the `hls.js` client synchronizes, and re-synchronizes, with the HLS stream every time, under all circumstances.  Without it, `hls.js` will not gain initial sync to a HLS stream if it is started just a few seconds after the stream has begun and, won't regain synchronization if it should lose it.

I did try running `hls.js` with `lowLatency: false` but that did not fix the problem.

Also, re-jiggered the web page to be more compact.