C++ code now obeying all of the configuration items, also scheduling its tasks in a real-time fashion nicely.
Gave up on luxon.js for the humanized display of time on the web interface as it doesn't really do it right (jumps when moving from hours to minutes left etc.), just not as good as moments.js, deprecated or not; now using moments.js.  Added highlighting of the current hour to the displayed schedule.

The software is now cooked aside from debugging of the LED behaviours; over to physical stuff.